requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
absl-py=1.4.0=pyhd8ed1ab_0
aiohttp=3.8.3=py39hb9d737c_1
aiosignal=1.3.1=pyhd8ed1ab_0
antlr4-python3-runtime=4.8=pypi_0
appdirs=1.4.4=pyh9f0ad1d_0
async-timeout=4.0.2=py39h06a4308_0
attrs=23.1.0=py39h06a4308_0
blas=1.0=mkl
blinker=1.5=pyhd8ed1ab_0
brotli=1.0.9=h166bdaf_8
brotli-bin=1.0.9=h166bdaf_8
brotlipy=0.7.0=py39hb9d737c_1005
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2023.12.12=h06a4308_0
cachetools=5.3.0=pyhd8ed1ab_0
casadi=3.6.4=pypi_0
certifi=2023.7.22=py39h06a4308_0
cffi=1.15.1=py39he91dace_3
charset-normalizer=2.1.1=pyhd8ed1ab_0
circle-fit=0.2.1=pypi_0
clarabel=0.6.0=pypi_0
click=8.1.3=unix_pyhd8ed1ab_2
cloudpickle=2.2.1=pypi_0
configargparse=1.7=pypi_0
connection-pool=0.0.3=pypi_0
contourpy=1.0.7=py39h4b4f3f3_0
control=0.9.3.post2=pyhd8ed1ab_0
cryptography=39.0.0=py39hd598818_0
cuda=11.7.1=0
cuda-cccl=11.7.91=0
cuda-command-line-tools=11.7.1=0
cuda-compiler=11.7.1=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cuobjdump=11.7.91=0
cuda-cupti=11.7.101=0
cuda-cuxxfilt=11.7.91=0
cuda-demo-suite=12.0.140=0
cuda-documentation=12.0.140=0
cuda-driver-dev=11.7.99=0
cuda-gdb=12.0.140=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=12.0.140=0
cuda-nsight-compute=12.0.1=0
cuda-nvcc=11.7.99=0
cuda-nvdisasm=12.0.140=0
cuda-nvml-dev=11.7.91=0
cuda-nvprof=12.0.146=0
cuda-nvprune=11.7.91=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-nvvp=12.0.146=0
cuda-runtime=11.7.1=0
cuda-sanitizer-api=12.0.140=0
cuda-toolkit=11.7.1=0
cuda-tools=11.7.1=0
cuda-visual-tools=11.7.1=0
cvxpy=1.4.1=pypi_0
cycler=0.11.0=pyhd8ed1ab_0
datrie=0.8.2=pypi_0
dbus=1.13.18=hb2f20db_0
docutils=0.20.1=pypi_0
dpath=2.1.6=pypi_0
ecos=2.0.12=py39h7deecbd_0
expat=2.5.0=h27087fc_0
fastjsonschema=2.19.1=pypi_0
ffmpeg=4.3=hf484d3e_0
flit-core=3.6.0=pyhd3eb1b0_0
fontconfig=2.14.1=hef1e5e3_0
fonttools=4.38.0=py39hb9d737c_1
freetype=2.10.4=h0708190_1
frozenlist=1.3.3=py39hb9d737c_0
gds-tools=1.5.1.14=0
giflib=5.2.1=h36c2ea0_2
gitdb=4.0.11=pypi_0
gitpython=3.1.40=pypi_0
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.16.0=pyh1a96a4e_1
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
grpcio=1.42.0=py39hce63b2e_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
gym=0.21.0=pypi_0
humanfriendly=10.0=pypi_0
hydra-core=1.1.0=pypi_0
hydra-joblib-launcher=1.2.0=pypi_0
hydra-submitit-launcher=1.1.5=pypi_0
icu=58.2=hf484d3e_1000
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.0.0=pyha770c72_0
intel-openmp=2021.4.0=h06a4308_3561
jinja2=3.1.2=py39h06a4308_0
joblib=1.2.0=pypi_0
jpeg=9e=h166bdaf_2
jsonschema=4.20.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-core=5.7.1=pypi_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=py39hf939315_1
krb5=1.19.3=h3790be6_0
lame=3.100=h7b6447c_0
lcms2=2.14=h6ed2654_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libblas=3.9.0=12_linux64_mkl
libbrotlicommon=1.0.9=h166bdaf_8
libbrotlidec=1.0.9=h166bdaf_8
libbrotlienc=1.0.9=h166bdaf_8
libcblas=3.9.0=12_linux64_mkl
libclang=10.0.1=default_hb85057a_2
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufft-dev=10.7.2.124=h98a8f43_0
libcufile=1.5.1.14=0
libcufile-dev=1.5.1.14=0
libcurand=10.3.1.124=0
libcurand-dev=10.3.1.124=0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20191231=he28a2e2_2
libevent=2.1.12=h8f2d780_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=12.2.0=h65d4601_19
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=12_linux64_mkl
libllvm10=10.0.1=he513fc3_3
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.15.8=h780b84a_1
libstdcxx-ng=12.2.0=h46fd767_19
libtasn1=4.16.0=h27cfd23_0
libtiff=4.4.0=hecacb30_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h522a892_0
libwebp-base=1.2.4=h166bdaf_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=hcb278e6_0
markdown=3.4.1=pyhd8ed1ab_0
markupsafe=2.1.3=py39h5eee18b_0
matplotlib=3.6.2=py39h06a4308_0
matplotlib-base=3.6.2=py39h945d387_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
mpmath=1.3.0=py39h06a4308_0
multidict=6.0.4=py39h72bdee0_0
munkres=1.1.4=pyh9f0ad1d_0
nbformat=5.9.2=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py39h06a4308_0
nsight-compute=2022.4.1.6=0
nspr=4.35=h27087fc_0
nss=3.74=h0370c37_0
numpy=1.24.2=py39h7360e5f_0
oauthlib=3.2.2=pyhd8ed1ab_0
omegaconf=2.1.2=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1w=h7f8727e_0
osqp=0.6.3=pypi_0
packaging=23.0=pyhd8ed1ab_0
pandas=2.2.0=pypi_0
pcre=8.45=h9c3ff4c_0
pd=0.0.4=pypi_0
pillow=9.3.0=py39h6a678d5_2
pip=22.3.1=py39h06a4308_0
plac=1.4.2=pypi_0
platformdirs=4.1.0=pypi_0
ply=3.11=py_1
pooch=1.6.0=pyhd8ed1ab_0
protobuf=3.15.8=py39he80948d_0
psutil=5.9.7=pypi_0
pulp=2.7.0=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pybind11=2.11.1=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pygame=2.5.2=pypi_0
pyglet=1.5.27=pypi_0
pyjwt=2.6.0=pyhd8ed1ab_0
pyopenssl=23.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.15.7=py39h6a678d5_1
pyqt5-sip=12.11.0=py39h6a678d5_1
pysocks=1.7.1=pyha2e5f31_6
python=3.9.16=h7a1cb2a_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.13.1=py3.9_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h67b0de4_1
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pypi_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=pypi_0
pyzmq=25.1.2=pypi_0
qdldl=0.1.7.post0=pypi_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
referencing=0.32.1=pypi_0
requests=2.28.2=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
reretry=0.11.8=pypi_0
rpds-py=0.16.2=pypi_0
rsa=4.9=pyhd8ed1ab_0
scipy=1.10.0=py39h7360e5f_2
scs=3.2.4.post1=pypi_0
setuptools=65.6.3=py39h06a4308_0
sip=6.6.2=py39h5a03fae_0
six=1.16.0=pyhd3eb1b0_1
slycot=0.5.3=py39ha6e88a2_0
smart-open=6.4.0=pypi_0
smmap=5.0.1=pypi_0
snakemake=7.32.4=pypi_0
sqlite=3.40.1=h5082296_0
stopit=1.1.2=pypi_0
submitit=1.4.5=pypi_0
tabulate=0.9.0=pypi_0
tensorboard=2.11.2=pyhd8ed1ab_0
tensorboard-data-server=0.6.1=py39hd97740a_4
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
termcolor=2.2.0=pyhd8ed1ab_0
throttler=1.2.2=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd8ed1ab_0
toposort=1.10=pypi_0
torchaudio=0.13.1=py39_cu117
torchvision=0.14.1=py39_cu117
tornado=6.2=py39hb9d737c_1
traitlets=5.14.1=pypi_0
typing_extensions=4.4.0=py39h06a4308_0
tzdata=2023.4=pypi_0
unicodedata2=15.0.0=py39hb9d737c_0
urllib3=1.26.14=pyhd8ed1ab_0
veins-gym=0.3.3=pypi_0
werkzeug=2.2.2=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.16.0=pypi_0
xz=5.2.10=h5eee18b_1
yarl=1.8.2=py39hb9d737c_0
yte=1.5.4=pypi_0
zipp=3.13.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zmq=0.0.0=pypi_0
zstd=1.5.2=ha4553b6_0
